Move table lazy loading from OcTd to OcTr
